# backend/storage/api-docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CMS Project API",
        "description": "    \n    ⚠️ Authorization Guide:\n    1. Use API **Login** to get Access Token.\n    2. Click on **[Authorize]** button.\n    3. Paste the token into the field.\n    4. If you receive **401**, it means you haven't logged in or the token has expired.\n    5. If you receive **403**, it means you don't have permission to access.\n\n    *Note: \n    - If you haven't logged in but the system still run the API (due to browser saving cookie), try again with anonymous.\n    - The System verify the token through cookie.\n    ",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Get role list",
                "description": "Display a listing of the resource.",
                "operationId": "fca5db0fa5666e4e2ebf340ec412bae3",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "currentPage",
                        "in": "query",
                        "description": "Current page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/RolesSearch"
                    },
                    "400": {
                        "$ref": "#/components/responses/Exception400"
                    },
                    "401": {
                        "$ref": "#/components/responses/Exception401"
                    },
                    "404": {
                        "$ref": "#/components/responses/Exception404"
                    },
                    "500": {
                        "$ref": "#/components/responses/Exception500"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Create role",
                "description": "Store a newly created resource in storage.",
                "operationId": "79a5495fb98e2354b8f7f35f2bc7406d",
                "requestBody": {
                    "$ref": "#/components/requestBodies/RolesCreate"
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/RolesCreate"
                    },
                    "400": {
                        "$ref": "#/components/responses/Exception400"
                    },
                    "401": {
                        "$ref": "#/components/responses/Exception401"
                    },
                    "404": {
                        "$ref": "#/components/responses/Exception404"
                    },
                    "500": {
                        "$ref": "#/components/responses/Exception500"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/roles/{id}": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Get role by id",
                "description": "Display the specified resource.",
                "operationId": "d819a9dc97f88b7506ce5e43af1d555f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Role id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/RolesGetById"
                    },
                    "400": {
                        "$ref": "#/components/responses/Exception400"
                    },
                    "401": {
                        "$ref": "#/components/responses/Exception401"
                    },
                    "404": {
                        "$ref": "#/components/responses/Exception404"
                    },
                    "500": {
                        "$ref": "#/components/responses/Exception500"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Roles"
                ],
                "summary": "Update role",
                "description": "Update the specified resource in storage.",
                "operationId": "931ddbfde151a20baabdb08ec40c02b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Role id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/RolesUpdate"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/RolesUpdate"
                    },
                    "400": {
                        "$ref": "#/components/responses/Exception400"
                    },
                    "401": {
                        "$ref": "#/components/responses/Exception401"
                    },
                    "404": {
                        "$ref": "#/components/responses/Exception404"
                    },
                    "500": {
                        "$ref": "#/components/responses/Exception500"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "Delete role",
                "description": "Remove the specified resource from storage.",
                "operationId": "44d383756460ee44e06eb294dffc1fd0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Role id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/RolesDelete"
                    },
                    "400": {
                        "$ref": "#/components/responses/Exception400"
                    },
                    "401": {
                        "$ref": "#/components/responses/Exception401"
                    },
                    "404": {
                        "$ref": "#/components/responses/Exception404"
                    },
                    "500": {
                        "$ref": "#/components/responses/Exception500"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Exception": {
                "required": [
                    "errorCode",
                    "errorMessage",
                    "data"
                ],
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "ERROR_CODE"
                    },
                    "message": {
                        "type": "string",
                        "example": "Error message."
                    },
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "example": []
                    }
                },
                "type": "object"
            },
...
            "Roles": {
                "required": [
                    "role_name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "uniqueItems": true,
                        "example": "admin"
                    },
                    "description": {
                        "type": "string",
                        "example": "web"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2023-01-01 00:00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2023-01-01 00:00:00"
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "example": []
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "Exception400": {
                "description": "Bad Request",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "$ref": "#/components/schemas/Exception"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "Exception401": {
                "description": "Unauthorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "$ref": "#/components/schemas/Exception"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "Exception404": {
                "description": "Not Found",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "$ref": "#/components/schemas/Exception"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "Exception500": {
                "description": "Internal Server Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "$ref": "#/components/schemas/Exception"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "RolesCreate": {
                "description": "OK",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "RolesDelete": {
                "description": "OK",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "RolesGetById": {
                "description": "OK",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "$ref": "#/components/schemas/Roles"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "RolesSearch": {
                "description": "OK",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Roles"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "RolesUpdate": {
                "description": "OK",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "$ref": "#/components/schemas/Roles"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "requestBodies": {
...
            "RolesCreate": {
                "description": "Create role",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "name",
                                "description"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "admin"
                                },
                                "description": {
                                    "type": "string",
                                    "example": "Admin role"
                                },
                                "permissions": {
                                    "type": "array",
                                    "items": {
                                        "type": "integer"
                                    },
                                    "example": [
                                        1,
                                        2,
                                        3
                                    ]
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "RolesUpdate": {
                "description": "Update role",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "name",
                                "description"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "admin"
                                },
                                "description": {
                                    "type": "string",
                                    "example": "Admin role"
                                },
                                "permissions": {
                                    "type": "array",
                                    "items": {
                                        "type": "integer"
                                    },
                                    "example": [
                                        1,
                                        2,
                                        3
                                    ]
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}